Display the times for the current day, for more consistency
ie display the sun rise time of the morning of the current day, not of the following day
(we can choose the following day if that's what we want)

also change a bit how to get it, by looking backwards from noon, so that it works
better close to the poles. Display the full date when a different day.